refactor: split EditorPage
EditorClient
- New custom hooks: useIsDark, useNativeAspectRatio, useSubtitles,
  useTextOverlays (+ useTextOverlayInspector/Interactions/Drag/Resize),
  useZoomEditor, useExportFlow, useSaveDemo, useDemoLoader, useAutosave,
  useEditorWiring, useEditorSyncEffects.
- New sub-components: EditorModals, EditorSidebarRegion, EditorPreviewRegion,
  EditorVideoStage, VideoPreviewPlayer, EmptyVideoState, SubtitleOverlay,
  ZoomFocusOverlay, TextOverlayLayer, TextOverlayItemView, TextColorToolbar,
  EditorTimelineSection.
- Pure helpers moved to module scope: resolveOverlayFontFamily (overlayFont),
  clamp, computeZoomPreview, exportHelpers, computePreviewLayout,
  notifyAutosavePending. Shared types in types.ts; api types in apiTypes.ts.
