+++ app/(signed)/editor/components/EditorModals.tsx
@@ -1,0 +1,107 @@
+import axios from "axios";
+import { Toaster } from "react-hot-toast";
+
+import ExportResultModal from "@/app/components/ExportResultModal";
+import ExportSettingsModal from "@/app/components/ExportSettingsModal";
+import SaveDemoModal from "@/app/components/SaveDemoModal";
+import ZoomEffectsPopup from "@/app/components/ZoomEffectsPopup";
+import type { EditorState, ExportFlowApi } from "../apiTypes";
+
+interface EditorModalsProps {
+  editorState: EditorState;
+  exportFlow: ExportFlowApi;
+  onSaveDemo: (data: { title: string; description: string }) => Promise<void>;
+  resolvedDuration: number;
+}
+
+export default function EditorModals({
+  editorState,
+  exportFlow,
+  onSaveDemo,
+  resolvedDuration,
+}: EditorModalsProps) {
+  const { pendingExport, shareLinkSaved, setShowExportResultModal, setPendingExport } = exportFlow;
+
+  return (
+    <>
+      <ExportSettingsModal
+        isOpen={exportFlow.showExportSettings}
+        onClose={() => exportFlow.setShowExportSettings(false)}
+        onConfirm={exportFlow.onExportVideo}
+        durationInSeconds={resolvedDuration || 0}
+      />
+      <ExportResultModal
+        isOpen={exportFlow.showExportResultModal}
+        shareUrl={pendingExport?.shareUrl || null}
+        loading={exportFlow.resultActionLoading}
+        onClose={() => {
+          if (exportFlow.resultActionLoading) {
+            return;
+          }
+
+          if (pendingExport && !shareLinkSaved) {
+            axios
+              .post("/api/exported-videos/cleanup", {
+                exportedUrl: pendingExport.exportedUrl,
+                sourceVideoUrl: pendingExport.uploadedSourceVideo
+                  ? pendingExport.sourceVideoUrl
+                  : null,
+                demoId: pendingExport.demoId,
+              })
+              .catch((e) => console.error("Cleanup on close failed:", e));
+          }
+          setShowExportResultModal(false);
+          setPendingExport(null);
+        }}
+        onDownloadMp4={exportFlow.handleDownloadMp4Only}
+        onSaveShareLink={exportFlow.handleSaveShareLink}
+      />
+      <Toaster
+        position="top-center"
+        toastOptions={{
+          className: "toast-popup-card",
+          style: {
+            background: "#2D2A3A",
+            color: "#fff",
+            fontWeight: "bold",
+            fontSize: "1.1rem",
+            boxShadow: "0 4px 24px 0 #0008",
+            borderRadius: "10px",
+            zIndex: 99999,
+          },
+          success: { iconTheme: { primary: "#7C5CFC", secondary: "#fff" } },
+          error: { iconTheme: { primary: "#f87171", secondary: "#fff" } },
+        }}
+      />
+
+      <ZoomEffectsPopup
+        isOpen={editorState.isZoomPopupOpen}
+        onClose={() => editorState.setIsZoomPopupOpen(false)}
+        zoomEffects={editorState.zoomEffects}
+        onZoomEffectsChange={editorState.setZoomEffects}
+        currentTime={editorState.currentTime}
+        duration={resolvedDuration}
+        onSeek={(time) => {
+          if (editorState.playerRef.current) {
+            const player = editorState.playerRef.current.getInternalPlayer();
+            if (player) {
+              player.currentTime = time;
+              player.dispatchEvent(new Event("seeking"));
+              editorState.playerRef.current.seekTo(time, "seconds");
+            }
+          }
+        }}
+      />
+
+      <SaveDemoModal
+        isOpen={editorState.showSaveDemoModal}
+        onClose={() => editorState.setShowSaveDemoModal(false)}
+        onSave={onSaveDemo}
+        initialTitle={editorState.sidebarTitle}
+        initialDescription={editorState.sidebarDescription}
+        processing={editorState.savingDemo}
+        isSaved={editorState.demoSaved}
+      />
+    </>
+  );
+}
