+++ app/(signed)/editor/utils/clamp.ts
@@ -1,0 +1,2 @@
+export const clamp = (value: number, min: number, max: number): number =>
+  Math.min(max, Math.max(min, value));
